+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added `is_candle_green()`, `is_candle_red()` helpers
 - Added `TimeBucket.to_pandas_timedelta()` helper
 - Added `upsample_candles()` helper
+- Added `unknown` and `osmosis` blockchains
  
 # 0.7